separated layers object and layers logic. put layers logic back in core.js. Also changed index to combine core.js and layers.js into single core.js in the build
