+++ app/templates/_index.html
@@ -56,11 +56,8 @@
     <!-- endbuild -->
 
     <!-- build:js scripts/core.js -->
-    <script src="scripts/core.js"></script>
-    <!-- endbuild -->
-
-    <!-- build:js scripts/layer.js -->
     <script src="scripts/layers.js"></script>
+    <script src="scripts/core.js"></script>
     <!-- endbuild -->
 
 </head>
