+++ app/templates/_main.css
@@ -52,8 +52,8 @@ body {
     /*height:320px;*/
     z-index:1;
     color: black;
-    background-color:white;
-    opacity:0.80;
+    background-color: rgb(255,255,255);
+    background-color: rgba(255,255,255,0.8);
     box-shadow: 0 1px 5px rgba(0,0,0,0.65);
     -webkit-border-radius: 4px;
     -moz-border-radius: 4px;
